feat: Extend joint analysis to include new averages
This commit enhances the exercise analysis capabilities by introducing the calculation of average angles for the following joints:
- Ankles
- Shoulders
- Elbows
- Wrists

These new metrics are calculated in the `analyze_frame` method and have been added to the `frame_data` dictionary. As a result, they are now part of the time series data, making them available to the frontend for more detailed user feedback.